==> documents/1-Fomatting data.qmd <==
---
title: "Formatting GLT data"
author: "Aurore Ponchon"
format: html
editor: visual
---

# 1- Formatting data to constitute a full long dataset

Data are stored in .xls or .xlsx files. They consist of blocks of data encompassing groups from the same region. A first formatting of the tables is made directly within the excel files to ensure that each group name is the first line of a block. Region names are isolated before blocks so that they are automatically assigned to following blocks.

```{r}

#| echo: false
#| message: false
#| warning: false

library(here)
library(plyr)
library(tidyverse)
library(lubridate)
library(readxl)
library(xlsx)
library(conflicted)


#Source files with specific functions
source(here::here("R","extract_functions.R"))
source(here::here("R","revalue_variables.R"))
source(here::here("R","progress_bar.R"))
source(here::here("R","extract_captures.R"))
source(here::here("R","extract_disp.R"))
source(here::here("R","extract_true_region_names.R"))

bad<-c("?","T0","IN","?-1","?-2","?-3","FT","?-4","?-5","?-6", "FT")


#identify particular status in Grupo column
dispersers<-c("OUT","Out","out","OUT PT",
              "disp","Disp","DISP",
              "Grace-disp","Alex-disp","28/01/13", 
              "saiu","Expulsa")
solo<-c("Sozinha","Sozinho","sozinha","sozinho",
        "SOZINHA","soz.","SOZINHO","Soz.")
dead<-c("DEAD","died","Died","DIED")
names<-c("chiquinho","Georgia","Luisa","Antônia","Luana",
         "Camilo","André")

#list all files
files.recent<-list.files(here::here("data","RawData","ObsData","2009-2022"),
                         pattern = "xls",recursive=T,
                         full.names = T)

files.old<-list.files(here::here("data","RawData","ObsData","2001-2009"),
                      pattern = "xls",recursive=T,
                      full.names = T)

filecapt<-here::here("data","RawData","ProcessamentoMLD.xlsx")
filedisp<-here::here("data","RawData","GLT_movement.xls")

#function to read each file separately
data.recent<- purrr::map_dfr(files.recent, extract_long_data,.progress=T)
data.pdf<- purrr::map_dfr(files.old, extract_long_data,.progress=T)

#bind observations
DATA<-rbind(data.pdf,data.recent)

#clean data with specific functions
data2000<-DATA %>% 
  revalue_region() %>% 
  revalue_name() %>% 
      revalue_birth() %>%    
  revalue_stage() %>% 
  revalue_sex() %>%
  dplyr::select(Year,DateObs,Region,Group,GLT,Tattoo,
                Sex,Birth,Idade, Disp, Death, Solo,File) 


#extract useful captures
datacapt<-extract_capture_data(filecapt) %>% 
  revalue_old_regions(.) %>% 
  dplyr::filter(Group %in% unique(data2000$Group)) %>% 
  revalue_names_capture()

datadisp<-extract_disp_data(filedisp) #disregard all the warnings

data.regions<-extract_true_region_names()


data.clean<-data2000 %>% 
  rbind(datacapt) %>%
  dplyr::mutate(DateObs=as.Date(DateObs,format="%Y-%m-%d"),
                Birth=as.Date(paste("01/",Birth, sep=""),format="%d/%m/%y")) %>% 
  rbind(datadisp) %>% 
  account_translocations() %>% 
  left_join(data.regions,by="Group", keep=F) %>% 
  
  dplyr::distinct(DateObs,Group,GLT,.keep_all = T) %>% 
   dplyr::arrange(DateObs,Group,GLT)


#check unique group names and Regions
# sort(unique(data.clean$Region))
# sort(unique(data.clean$Group))

#write a file with group and region names
# data.clean %>%
#   dplyr::filter(GLT!="?" &
#                   GLT !="?-3" &
#                   GLT!="IN" &
#                   GLT!="FT" &
#                   GLT!="T0" &
#                   Group!="?") %>%
#   dplyr::select(DateObs,Region, Group, 
#                 GLT, Sexo,Idade,Disp,Death,Solo) %>%

# write.table(., file=here::here("Data","NewlyCreatedData","data_long.xslx"),
#            quote=F,row.names=F ,sep=";")

save(data.clean,file=here::here("Data",
                                "NewlyCreatedData",
                                "clean_raw_data_long.RData"))

```

# 2- Checking duplicates and errors

Duplicates are checked in group and region names, as well as in individuals. Individual variables are also filled, to ensure that they are consistent among observations of the same individual (notably sex and date birth).

```{r}
#| echo: false
#| message: false
#| warning: false
source(here::here("R","GLT_romain_data_correction.R"))
data.clean.final<-correcting_sex_stage() 

#filter only necessary data
data.clean.final<-data.clean.final%>% 
  dplyr::select(Year,DateObs,DateObsMY,Region,UMMPs,
                Group,GLT,Tattoo,SexOK, BirthOK,IdadeOK,
                Disp,Death,Solo,
                File,season,ObsOrder) %>% 
   as.data.frame(.)

## Export the corrected dataset
 save(data.clean.final, file=here::here("data","NewlyCreatedData","data_clean_long_final.RData"))

# #identify duplicates in individual names
# #create an object with GLT duplicates
# dup<-data.clean %>%
#   dplyr::filter(!GLT %in% bad) %>%
#   dplyr::filter(Disp==0) %>% 
#   dplyr::group_by(DateObs,GLT) %>%
#   dplyr::summarise(n = dplyr::n(), .groups = "drop") %>%
#   dplyr::filter(n > 1L) # & GLT!="?" & GLT!="T0" & GLT!="IN") %>% 
# 
# #get all observations of duplicates to reconstitute their individual histories
# 
# dups<-data.clean %>% 
#   dplyr::filter(GLT %in% dup$GLT) %>% 
#   dplyr::arrange(GLT, DateObs)
# 
# ###
# ##Individual Cibele wuth different sex
# #RL15, AF6, Af8
# 
# ##checking duplicates in Sex
# sexo<-data.clean %>%
#   dplyr::filter(Sex!="nonID") %>% 
#   dplyr::group_by(GLT) %>%
#   dplyr::summarise(Sexy = Sex,.groups="keep") %>% 
#   distinct() %>% 
#   dplyr::filter(n()>1)
# 
# write.table(sexo,file=here::here("data","NewlyCreatedData","Checks","sex_check.txt"),
#       sep="\t",row.names=F,quote=F)
# 
# #MP475 problematic (same birth and both M & F)
# 
# #1375 with 3 different dates!!
# #AF22, AF23, AF24, AF25 born after observation and dispersed?!
# #check FA40 (2018 & 2019)
# #JP23 duplicate
# #check 1311, 1312, MP104, MP108, MP109, MP176, MP292, MP313, MP99, TM20, UR4
# 
# #check differences between GLT and tattoo
# ind<-data.clean %>% 
#   dplyr::filter(!is.na(Tattoo)) %>% 
#   dplyr::filter(GLT!=Tattoo) %>% 
#   dplyr::arrange(GLT,DateObs)
# 
# 
# #check duplicates with groups and regions
# group<-data.clean %>% 
#   dplyr::filter(Disp!=1 & Death==0 & Solo ==0) %>% 
#   dplyr::group_by(Group) %>% 
#   dplyr::select(Group,Region) %>% 
#   dplyr::distinct() %>% 
#   dplyr::arrange(Group) %>% 
#   as.data.frame(.) %>% 
#   write.xlsx(., file=here::here("data","NewlyCreatedData","GroupRegionCheck.xlsx"),
#                           col.names=T,
#              row.names=F)



```

# 3- Checking continuity in observations

```{r}
#| echo: false
#| message: false
#| warning: false

library(viridis)
library(gridExtra)

source(here::here("R","plot_continuity.R"))

# fq<-data.clean.final %>%  
#   dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0 & Solo==0) %>% 
#   group_by(UMMPs,Group,DateObs) %>% 
#   dplyr::count() %>% 
#    group_by(UMMPs,Group) %>% 
#    dplyr::count() %>% 
#   dplyr::arrange(UMMPs,n) %>% 
#   drop_na()
# 
# colorme<-rep(viridis(n=22,option="H"),time=11)
# title_plot<-"Monitoring continuity in all regions"
# title_file<-"Continuity_field_effort_all_regions.tif"
# 
#  
# plot_continuity(data.clean.final,fq,title_plot,title_file,W=6000,H=6500)
#   
#   
#    regio<-sort(unique(data.clean.final$UMMPs))
#    
#    
#    for (k in 1:length(regio)){
#      sub<-data.clean.final %>% 
#        dplyr::filter(UMMPs==regio[k]) 
#        
#       fq<-sub %>%  
#   group_by(Group) %>% 
#   dplyr::count() %>% 
#   dplyr::arrange(n) %>% 
#   drop_na()
#       
# 
# title_plot<-paste("Monitoring continuity in ",
#                   regio[k],
#                   sep="")
# title_file<-paste("Continuity_field_effort_",
#                   regio[k],
#                   ".tif",
#                   sep="") 
#     
#  plot_continuity(sub,fq,title_plot,title_file,W=6000,H=6500)   
#  setTxtProgressBar(progress_bar(length(regio)),k)  
#  
#    }
   
   #Nombre d'observations par groupe
   
   loc<-read.table(here::here("data","rawData","Landscape","RegionsName.csv"),
                header=T,sep=";") %>% 
  dplyr::rename(Long=CENTROIDE.X.UTM.SAD69.23S,
                Lat=CENTROIDE.Y.UTM.SAD69.23S,
                Group=Abreviation) %>% 
  dplyr::select(-Platform,-City) %>% 
  dplyr::filter_all(any_vars(!is.na(.)))
   
   
   
   
  gpobs<- data.clean.final %>%  
  dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0 & Solo==0) %>% 
  group_by(Group) %>% 
   dplyr::summarize(MinYear=min(Year),
                    MaxYear=max(Year),
                    Duration=max(Year)-min(Year),
                    N=n_distinct(DateObs), 
                    UMMPs=unique(UMMPs)[1],
                    .groups = "drop") %>% 
     dplyr::arrange(desc(N)) %>% 
    left_join(loc) %>% 
    
     write.csv2(here::here("data","NewlyCreatedData","Nb_obs_gp.csv"),
                row.names=F)
     
  
  ##Figure supplemental material to show monitoring effort
  
  
  library(viridis)
library(gridExtra)

source(here::here("R","plot_continuity.R"))

fq<-data.clean.final %>%  
  dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0 & Solo==0) %>% 
  group_by(UMMPs,Group,DateObs) %>% 
  dplyr::count() %>% 
  group_by(UMMPs,Group) %>% 
  dplyr::count() %>% 
  dplyr::arrange(UMMPs,n) %>% 
  drop_na()

colorme<-viridis(n=length(unique(data.clean.final$UMMPs))-1,option="H")
colo_df<-data.frame(Xcolor=colorme,UMMPs=sort(unique(data.clean.final$UMMPs)))

title_plot<-"Monitoring effort in all forest fragments"
title_file<-"Field_effort_all_regions_1989-2022.tif"


check<-data.clean.final %>% 
  dplyr::filter(!GLT %in% bad & Disp==0 & Death==0 & Solo==0) %>%
  dplyr::mutate(Group=as.factor(Group),
                DateObs=as.Date(DateObs, format="%F")) %>%
    dplyr::mutate(Group=factor(Group,
                               levels=fq$Group)) %>% 
  dplyr::filter(!is.na(Group)) %>% 
  dplyr::arrange(UMMPs) %>%  
  left_join(.,colo_df,by="UMMPs")


colo.nb<-check %>%   
  dplyr::group_by(Group,UMMPs,Xcolor) %>% 
  dplyr::slice(1) %>% 
  dplyr::select(Group,UMMPs,Xcolor) 
  
 check.plot<- ggplot(check,aes(x=DateObs,
               y=Group, colour=UMMPs)) +
  geom_point( shape=15,
             size=1.2) +
 scale_colour_manual(values=colorme, name="Forest fragments") +
  scale_y_discrete(guide=guide_axis(n.dodge = 2),
                   labels = function(x) {
                     col <- colo.nb$Xcolor
                     lab <- colo.nb$Group
                     glue::glue("<span style = 'color:{col}'>{lab}</span>")}) + 
  scale_x_date(limits=c(as.Date("1989-01-01",format="%F"),
                        as.Date("2024-01-01",format="%F")),
               breaks=seq(as.Date("1989-01-01",format="%F"),
                          as.Date("2024-11-01",format="%F"),
                          "5 years"),
               labels=c("1989","1994","1999","2004","2009",
                        "2014","2019","2024")) +
  theme_bw() +
  labs(x="Time",
       y="Groups",
       title=title_plot) +
  theme( axis.text.y = element_markdown())

print(check.plot)


# if(nrow(fq) > 50){
#   
# }
# 
tiff(here::here("outputs","Monitoring effort",title_file),
     width=W,
     height=H,
     res=600,
     compression="lzw")


print(check)
dev.off()

col="#30123BFF"

  
```

# 4- Identifying successful dispersal events

```{r}

#| echo: false
#| message: false
#| warning: false

#filter data and add id count over time
data.filter<-data.clean.final %>% 
  dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0) %>% 
  dplyr::arrange(DateObs, GLT) %>% 
  group_by(GLT) 
  
#get the name of all individuals observed in more than one group
dispind<-data.filter %>%
  dplyr::group_by(GLT) %>%
  dplyr::summarise(N = n_distinct(Group)) %>% 
  dplyr::filter(N >1)

data.disp<-data.filter %>% 
  dplyr::filter(GLT %in% dispind$GLT) %>% 
  dplyr::arrange(GLT,DateObs) %>% 
  dplyr::mutate(FromGroup=dplyr::lag(Group),
                ToGroup=Group,
                FromRegion=dplyr::lag(Region),
                ToRegion=Region,
                FromUMMP=dplyr::lag(UMMPs),
                ToUMMP=UMMPs) %>% 
  dplyr::mutate(Transloc=ifelse(dplyr::lag(Disp==2) & 
                                  Disp==0 & 
                                  dplyr::lag(GLT)==GLT,
                                1,0)) %>% 
  dplyr::filter(ObsOrder>1) %>%    #remove first observation
  
  dplyr::mutate(Change=ifelse(ToGroup != FromGroup, 1,0)) %>% 
   dplyr::filter(Change > 0 & Transloc == 0) %>% 
   dplyr::select(Year,DateObs,GLT,Tattoo,SexOK,BirthOK,IdadeOK,
                 FromGroup,ToGroup,FromRegion,ToRegion,FromUMMP,ToUMMP)


dispind[which(dispind$N>2),]
  

  write.xlsx(as.data.frame(data.disp),
             file=here::here("data","NewlyCreatedData","Successful_dispersal.xlsx"),
             sheetName = "Sheet1",
             col.names=T,
             row.names=F)



```

# 5- Identifying all dispersal events

```{r}
#| echo: false
#| message: false
#| warning: false


#filter data and add id count over time
data.filter<-data.clean.final %>% 
  dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0) %>% 
  dplyr::arrange(DateObs, GLT) %>% 
  group_by(GLT) 

disp<-NULL

gp<-sort(unique(data.filter$Group))

for (i in 1:length(gp)){
  
singlegp<-data.filter %>% 
  dplyr::filter(Group==gp[i])

dateobs<-sort(unique(singlegp$DateObs))

if(length(dateobs)>1){
for (j in 2:length(dateobs)){
  
  daygpbef<-singlegp %>% 
    dplyr::filter(DateObs==dateobs[j-1])
  
  daygp<-singlegp %>% 
    dplyr::filter(DateObs==dateobs[j])
  
  miss<-anti_join(daygpbef,daygp,by="GLT")
  
  if(nrow(miss)>0){
disp<-rbind(disp,miss)

  }
  setTxtProgressBar(progress_bar(length(gp)),i)  
}
}
}

emigr<-disp %>% 
  group_by(GLT) %>% 
  dplyr::slice(max(ObsOrder)) %>% 
  as.data.frame(.) %>% 
  
  
  write.xlsx(.,file=here::here("data","NewlyCreatedData",
                               "All_dispersal_events_lastobs.xlsx"),
             sheetName = "Sheet1",
             col.names=T,
             row.names=F)






```
