==> documents/3- Linking locations with groups.qmd <==
---
title: "Joining groups with spatial locations"
author: "Aurore Ponchon"
format: html
editor: visual
---

# Joining groups with spatial locations

Groups are assigned to a region which is itself located with projected coordinates in the local EPSG 31983 projection system

```{r}
library(here)
library(tidyverse)
library(sf)
library(ggspatial)
library(ggrepel)

source(here::here("R","progress_bar.R"))

load(here::here("Data","NewlyCreatedData","data_clean_long_final.RData"))
bad<-c("?","T0","IN","?-1","?-2","?-3","FT","?-4","?-5","?-6", "FT")


#Import file with locations associated with groups
loc<-read.table(here::here("data","rawData","Landscape","RegionsName.csv"),
                header=T,sep=";") %>% 
  dplyr::rename(Long=CENTROIDE.X.UTM.SAD69.23S,
                Lat=CENTROIDE.Y.UTM.SAD69.23S,
                Group=Abreviation) %>% 
  dplyr::select(-Platform,-City) %>% 
  dplyr::filter_all(any_vars(!is.na(.)))
  

#Join groups and locations for individual observations
excl<-c("AF","JG", "PA")


spatdata<-data.clean.final %>% 
   dplyr::left_join(.,loc, by="Group",
            multiple="first") %>% 
  dplyr::filter(!is.na(Long)) %>%
  dplyr::filter(!Group %in% excl) %>% 
  dplyr::filter(Year > 2009) %>% 
  droplevels() %>% 
  dplyr::arrange(GLT,Group,DateObs) %>% 
  dplyr::mutate(GLT=as.factor(GLT)) %>% 
  dplyr::group_by(GLT)%>% 
  dplyr::filter(!GLT %in% bad & Disp!=1 & Death==0 & Solo==0) %>% 
sf::st_as_sf(., coords = c("Long","Lat")) %>% 
  sf::st_set_crs(31983) 


loc.groups<-spatdata %>% 
  group_by(Group) %>% 
  distinct(UMMPs,geometry) %>% 
  write_sf(.,
         here::here("data","NewlyCreatedData","Dispersal lines",
                    "group_locs_filtered.shp"),
                    delete_dsn = T,
                    delete_layer=T)



land<-sf::read_sf(here::here("data","RawData","Landscape","Shapefiles Landscape AMLD", "SIG-EDUC Redescobrindo 2021 - Fragmentos de Vegetação.shp")) 

ummp<-sf::read_sf(here::here("data","RawData","Landscape","Shapefiles Landscape AMLD", "SIG-LGCI_UMMP-13.shp")) 



region_groups<-st_join(spatdata,land)

#Identify changes in groups =dispersal)
ids<-unique(spatdata$GLT)
spatdisp<-NULL

for (i in 1:length(ids)) {
  temp<-subset(spatdata,spatdata$GLT==ids[i] & spatdata$Disp!=2)
  temp$disp<-0
  if(nrow(temp)>1){
    for (j in 2:nrow(temp)){
      if(temp$Group[j-1]!=temp$Group[j]){
        temp$disp[c(j-1,j)]<-1
      }
    }
    spatdisp<-rbind(spatdisp,temp)   
     setTxtProgressBar(progress_bar(length(ids)),i)  
  }
}  

spatdisp1<-spatdisp %>% 
  dplyr::filter(disp==1) %>% 
 # distinct(GLT,DateObs, .keep_all = TRUE) %>% 
  dplyr::arrange(Year,GLT) 

spdisp<-spatdisp1 %>% 
  sf::st_as_sf(., coords = c("Long","Lat")) %>% 
  sf::st_set_crs(31983) %>% 
  dplyr::mutate(Year=as.factor(Year))

lines_disp <- spdisp %>% 
  group_by(GLT) %>%
  dplyr::summarise(do_union = FALSE) %>%
  st_cast("LINESTRING") 


lines_disp<- lines_disp %>% 
  dplyr::mutate(Dist=as.vector(st_length(lines_disp,by="id")/1000))

write_sf(lines_disp,
         here::here("data","NewlyCreatedData","Dispersal lines",
                    "disp_lines_2010.shp"),
                    delete_dsn = T,
                    delete_layer=T)


#jjit<-st_jitter(spdisp, factor = 0.1)

fieldgg<-ggplot()+
  geom_sf(data = land,
          colour="grey70",
          alpha=0.7) + 
  geom_sf(data=ummp,
          aes(fill=UMMPs),
          alpha=0.5) +
  geom_sf(data=lines_disp,
          size=0.2) +  
  geom_sf(data=spdisp,
          aes(colour=Group),
          size=0.5) +  
  
  # geom_sf_text(data=spdisp,
  #              mapping=aes(label = Group),
  #              nudge_x=100,
  #              nudge_y=50) +
   annotation_scale(location = "br",
                    width_hint = 0.3,
                    pad_x = unit(0.3, "cm"),
                   bar_cols = "black") +
   annotation_north_arrow(location = "bl",
                         which_north = "true",
                          style = ggspatial::north_arrow_nautical(),
                          pad_y = unit(0.2, "cm")) +
  theme_void() +
  theme(legend.position = "none")



tiff(filename=here::here("outputs","Dispersal_lines_2010.tif"), 
            res=300,
            units="px",
            height=1500,
            width=3000,
            compression="lzw")
print(fieldgg)
dev.off()
```
