==> documents/4-Landscape analysis on metrics.qmd <==
---
title: "Landscape analysis"
author: "Aurore Ponchon"
format: html
editor: visual
---

# Linking landscape composition with individual movements

```{r}

library(here)
library(raster)
library(tidyverse)
library(sf)
library(landscapemetrics) #package specific to landscape analysis
source(here::here("R","progress_bar.R"))
source(here::here("R","create_raster.R"))
source(here::here("R","merge_ummp.R"))
#import landscape files

#raster land cover for each year obtained from BioMAPS
rast<-list.files(here::here("data","RawData","Landscape","Rasters land cover"),
                 full.names=T)

#forest patches from shapefile
land<-sf::read_sf(here::here("data","RawData","Landscape","Shapefiles Landscape AMLD", 
                             "SIG-EDUC Redescobrindo 2021 - Fragmentos de Vegetação.shp"))


#conservation units (UMMPs)
ummp<-return_complete_ummp() #add vendaval and boa esperanza which are missing in shapefile
  
#remove accents in names
ummp$UMMPs<-plyr::revalue(ummp$UMMPs, c("Imbaú I" = "Imbau I",
                         "Imbaú II" = "Imbau II",
                         "Imbaú III" = "Imbau III",
                         "Poço das Antas" = "Poco das Antas",
                         "União I" = "Uniao I",
                         "União II" = "Uniao II"))

#get locations of groups and set them as shape layer
loc<-read.table(here::here("data","rawData","Landscape","RegionsName.csv"),
                header=T,sep=";") %>% 
  dplyr::rename(Long=CENTROIDE.X.UTM.SAD69.23S,
                Lat=CENTROIDE.Y.UTM.SAD69.23S,
                Group=Abreviation) %>% 
  dplyr::select(-Platform,-City) %>% 
  dplyr::filter_all(any_vars(!is.na(.))) %>% 
  sf::st_as_sf(., coords = c("Long","Lat")) %>% 
  sf::st_set_crs(31983)

#extent of study area in geographic coordinates
extreg<-extent(-42.7,-41.9,-22.9,-22.3)
size<-NULL

#projection for Brazil
#proj<-"+proj=utm +zone=23 +south +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs +type=crs"
proj<-"EPSG:31983"
#get all annual land cover in the same raster stack
 #coverStack<-purrr::map(rast,create_raster_stack,.progress=T)
 #save(coverStack, file=here::here("data","NewlyCreatedData","Cover_rasters.RData"))

 load(here::here("data","NewlyCreatedData","Landscape raster", "Rasters",
                 "Cover_rasters.RData"))

#create an initial raster for size area
rastsize<-create_raster_stack(rast[1])

#create an initial raster for patches identified with package landscapemetrics
patches<-create_raster_stack(rast[1])

#create an initial raster for patch perimeter
rastperim<-create_raster_stack(rast[1])

#create an initial raster for patch shape
rastshape<-create_raster_stack(rast[1])

#loop through each annual land cover
for (i in 1:length(rast)){
# i<-2

#get land cover
cover1<-create_raster_stack(rast[i])

setTxtProgressBar(progress_bar(length(rast)),i)  

#assign habitat types
#check_landscape(rast1)
#show_patches(rast1,class="all")

#get a raster with patch id 
zz<- get_patches(cover1,class=1)
zzz<-zz$layer_1$class_1
names(zzz)<-paste("Patches",parse_number(rast[i]),sep="")
patches<-stack(zzz,patches)


#get a raster for SIZE AREA
xx<-spatialize_lsm(cover1, what = "lsm_p_area")
rr<-xx$layer_1$lsm_p_area
names(rr)<-paste("Area",parse_number(rast[i]),sep="")
 rastsize<-stack(rr,rastsize)
 
# get a raster for PERIMETER
pp<-spatialize_lsm(cover1, what = "lsm_p_perim")
pp<-pp$layer_1$lsm_p_perim
names(pp)<-paste("Perimetre",parse_number(rast[i]),sep="")
 rastperim<-stack(pp,rastperim)

#get a raster for  SHAPE
sh<-spatialize_lsm(cover1, what = "lsm_p_shape")
sh<-sh$layer_1$lsm_p_shape
names(sh)<-paste("Shape",parse_number(rast[i]),sep="")
 rastshape<-stack(sh,rastshape)
 
 #extract all possible variable that can be calculated by landscapemetric
# yy<-extract_lsm(cover1,loc)  %>% 
#    dplyr:: mutate(Year=paste("cover",readr::parse_number(rast[i]),sep=""))
  # size<-rbind(size,yy)
 
}


# terra::writeRaster(rasthabitat,
#                     filename=here::here(paste0("data",
#                                        "NewlyCreatedData",
#                                        "Landscape raster",
#                                        names(rasthabitat))),
#                     overwrite=T)
# 
# terra::writeRaster(patches,
#                     filename=here::here(paste0("data",
#                                        "NewlyCreatedData",
#                                        "Landscape rast")),
#                     overwrite=T,
#                     bylayer=T)


save(size,file=here::here("data","NewlyCreatedData","SizeArea_table.RData"))
save(rastsize,file=here::here("data","NewlyCreatedData","Raster_size.RData"))
save(patches,file=here::here("data","NewlyCreatedData","Raster_patches.RData"))
save(rastperim,file=here::here("data","NewlyCreatedData","Raster_perim.RData"))
save(rastshape,file=here::here("data","NewlyCreatedData","Raster_shape.RData"))

```

# GIF of habitat cover and size area

```{r}

library(gganimate)
library(gifski)

load(here::here("data","NewlyCreatedData","data_clean_long_final.RData"))
load(here::here("data","NewlyCreatedData","Cover_rasters.RData"))
load(here::here("data","NewlyCreatedData","Raster_size.RData"))
load(here::here("data","NewlyCreatedData","Raster_perim.RData"))
load(here::here("data","NewlyCreatedData","Raster_shape.RData"))
load(here::here("data","NewlyCreatedData","Raster_patches.RData"))


#get regions with groups
group.reg<-data.clean.final %>% 
  ungroup() %>% 
  distinct_at(vars(Year,Group,UMMPs), keep_all=F)

reg<-sort(unique(group.reg$UMMPs))

totsizearea<-NULL
totshape<-NULL
totperim<-NULL
patchNB<-NULL

#loop trhough each UMMP region
for (p in 1:length(reg)){
  
  setTxtProgressBar(progress_bar(length(reg)),p)  

pda<- ummp %>% 
   dplyr::filter(UMMPs==reg[p])

#crop landscape to given UMMP
habitatcrop<- purrr::map(coverStack,crop,pda)
habcropstack<-stack(habitatcrop)

#crop all rasters to region
sizecrop<-crop(rastsize,pda)
perimcrop<-crop(rastperim,pda)
shapecrop<-crop(rastshape,pda)
patchcrop<-crop(patches,pda)

#extract habitat code for each group location
habarea<-terra::extract(habcropstack,loc) %>% 
  as.data.frame(.) %>% 
   cbind(loc) %>% 
#  dplyr::filter(Farm=="Rebio PDA" | Farm=="PDA") %>% 
  left_join (group.reg, by="Group") %>% 
  pivot_longer(cols=1:32,
               names_to="Cover_year",
               values_to="Habitat") %>% 
  dplyr::mutate(Year=parse_number(Cover_year),
                Fragment.Region=as.factor(Fragment.Region),
                Group=as.factor(Group),
   #             Patch= as.factor(brasil_coverage_1990),
                UMMPs=as.factor(UMMPs)) %>% 
# dplyr::filter(UMMPs=="Poco das Antas") %>%    #Aldeia II,  Rio Vermelho
  dplyr::arrange(Group,Year) %>% 
  drop_na() %>% 
  distinct()

#extract habitat code for each group location
patchnb<-terra::extract(patchcrop,loc) %>% 
  as.data.frame(.) %>% 
   cbind(loc) %>% 
#  dplyr::filter(Farm=="Rebio PDA" | Farm=="PDA") %>% 
  left_join (group.reg, by="Group") %>% 
  pivot_longer(cols=1:32,
               names_to="Cover_year",
               values_to="Habitat") %>% 
  dplyr::mutate(Year=parse_number(Cover_year),
                Fragment.Region=as.factor(Fragment.Region),
                Group=as.factor(Group),
   #             Patch= as.factor(brasil_coverage_1990),
                UMMPs=as.factor(UMMPs)) %>% 
# dplyr::filter(UMMPs=="Poco das Antas") %>%    #Aldeia II,  Rio Vermelho
  dplyr::arrange(Group,Year) %>% 
  drop_na() %>% 
  distinct()

patchNB<-rbind(patchNB,patchnb)

#extract patch size for each group location
sizearea<-terra::extract(sizecrop,loc) %>% 
  as.data.frame(.) %>% 
   cbind(loc) %>% 
#  dplyr::filter(Farm=="Rebio PDA" | Farm=="PDA") %>% 
  left_join (group.reg, by="Group") %>% 
  pivot_longer(cols=1:32,
               names_to="Cover_year",
               values_to="Size") %>% 
  dplyr::mutate(Year=parse_number(Cover_year),
                Fragment.Region=as.factor(Fragment.Region),
                Group=as.factor(Group),
   #             Patch= as.factor(brasil_coverage_1990),
                UMMPs=as.factor(UMMPs)) %>% 
# dplyr::filter(UMMPs=="Poco das Antas") %>%    #Aldeia II,  Rio Vermelho
  dplyr::arrange(Group,Year) %>% 
  drop_na() %>% 
  distinct()

totsizearea<-rbind(totsizearea,sizearea)


#perimeter length
perimarea<-terra::extract(perimcrop,loc) %>% 
  as.data.frame(.) %>% 
   cbind(loc) %>% 
#  dplyr::filter(Farm=="Rebio PDA" | Farm=="PDA") %>% 
  left_join (group.reg, by="Group") %>% 
  pivot_longer(cols=1:32,
               names_to="Cover_year",
               values_to="Perimetre") %>% 
  dplyr::mutate(Year=parse_number(Cover_year),
                Fragment.Region=as.factor(Fragment.Region),
                Group=as.factor(Group),
   #             Patch= as.factor(brasil_coverage_1990),
                UMMPs=as.factor(UMMPs)) %>% 
# dplyr::filter(UMMPs=="Poco das Antas") %>%    #Aldeia II,  Rio Vermelho
  dplyr::arrange(Group,Year) %>% 
  drop_na() %>% 
  distinct()

totperim<-rbind(totperim,perimarea)

#perimeter length
shapearea<-terra::extract(shapecrop,loc) %>% 
  as.data.frame(.) %>% 
   cbind(loc) %>% 
#  dplyr::filter(Farm=="Rebio PDA" | Farm=="PDA") %>% 
  left_join (group.reg, by="Group") %>% 
  pivot_longer(cols=1:32,
               names_to="Cover_year",
               values_to="Shape") %>% 
  dplyr::mutate(Year=parse_number(Cover_year),
                Fragment.Region=as.factor(Fragment.Region),
                Group=as.factor(Group),
   #             Patch= as.factor(brasil_coverage_1990),
                UMMPs=as.factor(UMMPs)) %>% 
# dplyr::filter(UMMPs=="Poco das Antas") %>%    #Aldeia II,  Rio Vermelho
  dplyr::arrange(Group,Year) %>% 
  drop_na() %>% 
  distinct()

totshape<-rbind(totshape,shapearea)

#convert raster stack to dataframe
habrast<- purrr::map_dfr(
  habitatcrop, 
  ~setNames(as.data.frame(as(., "SpatialPixelsDataFrame")),
            c('value', 'x', 'y')), 
            .id = 'year') %>% 
  dplyr::mutate(Year=as.numeric(year) + 1989)


sizerast<- sizecrop %>% 
  as.list(.) %>% 
  purrr::map_dfr(., 
  ~setNames(as.data.frame(as(., "SpatialPixelsDataFrame")),
            c('value', 'x', 'y')), 
            .id = 'year') %>% 
  dplyr::mutate(Year=as.numeric(year) + 1989)

shaperast<- shapecrop %>% 
  as.list(.) %>% 
  purrr::map_dfr(., 
  ~setNames(as.data.frame(as(., "SpatialPixelsDataFrame")),
            c('value', 'x', 'y')), 
            .id = 'year') %>% 
  dplyr::mutate(Year=as.numeric(year) + 1989)


perimrast<- perimcrop %>% 
  as.list(.) %>% 
  purrr::map_dfr(., 
  ~setNames(as.data.frame(as(., "SpatialPixelsDataFrame")),
            c('value', 'x', 'y')), 
            .id = 'year') %>% 
  dplyr::mutate(Year=as.numeric(year) + 1989)


#get coordinates in a dataframe
groupy<-sizearea %>% 
   mutate(long = unlist(map(sizearea$geometry,1)),
           lat = unlist(map(sizearea$geometry,2))) %>% 
  dplyr::select(Year,UMMPs,Size,long,lat,Group)  %>% 
  dplyr::filter(UMMPs==reg[p])


# gghab <- ggplot(habrast, aes(x = x, y = y)) +
#   geom_tile(aes(fill = value)) +
#   scale_fill_viridis_c()  +
#    geom_point(data=groupy, aes(x=long,y=lat,color=Group),shape=17,size=4,
#               show.legend=F)
# print(gghab)
# 
# gganimhab <- gghab +
#   transition_time(as.numeric(Year)) +
#    labs(subtitle = "Year: {round(frame_time,digits=0)}")
# 
# print(c("animate habitat ", p))
# gganimate::animate(gganimhab, height = 1000, width =2000,nframes=32,res=250,
#         renderer = gifski_renderer())
# anim_save(filename = here::here("outputs", "GIF cover",
#                                 paste("GIF_cover_",reg[p],".gif")))


# ggsize <- ggplot(sizerast, aes(x = x, y = y)) +
#   geom_tile(aes(fill = value)) +
#   scale_fill_viridis_c()  +
#    geom_point(data=groupy, aes(x=long,y=lat,color=Group),shape=17,size=4,
#               show.legend=F)
# print(ggsize)
# 
# gganimsize <- ggsize +
#   transition_time(as.numeric(Year)) +
#    labs(subtitle = "Year: {round(frame_time,digits=0)}")
# 
#  print(c("animate size ", p))
# 
# gganimate::animate(gganimsize, height = 1000, width =2000,nframes=32,res=250,
#         renderer = gifski_renderer())
# anim_save(filename = here::here("outputs","GIF cover",paste0("GIF_size_",reg[p],
#                                        ".gif")))


# ggshape <- ggplot(shaperast, aes(x = x, y = y)) +
#   geom_tile(aes(fill = value)) +
#   scale_fill_viridis_c()  +
#    geom_point(data=groupy, aes(x=long,y=lat,color=Group),shape=17,size=4,
#               show.legend=F)
# print(ggshape)
# 
# gganimshape <- ggshape +
#   transition_time(as.numeric(Year)) +
#    labs(subtitle = "Year: {round(frame_time,digits=0)}")
# 
#  print(c("animate shape ", p))
# 
# gganimate::animate(gganimshape, height = 1000, width =2000,nframes=32,res=250,
#         renderer = gifski_renderer())
# anim_save(filename = here::here("outputs","GIF cover",paste0("GIF_shape_",reg[p],
#                                        ".gif")))


# ggperim <- ggplot(perimrast, aes(x = x, y = y)) +
#   geom_tile(aes(fill = value)) +
#   scale_fill_viridis_c()  +
#    geom_point(data=groupy, aes(x=long,y=lat,color=Group),shape=17,size=4,
#               show.legend=F)
# print(ggperim)
# 
# gganimperim <- ggperim +
#   transition_time(as.numeric(Year)) +
#    labs(subtitle = "Year: {round(frame_time,digits=0)}")
# 
#  print(c("animate perimeter ", p))
# 
# gganimate::animate(gganimperim, height = 1000, width =2000,nframes=32,res=250,
#         renderer = gifski_renderer())
# anim_save(filename = here::here("outputs","GIF cover",paste0("GIF_perimetre_",reg[p],
#                                        ".gif")))
}

save(totsizearea,file=here::here("data","NewlyCreatedData","TotalSizeArea.RData"))
save(totshape,file=here::here("data","NewlyCreatedData","TotalShape.RData"))
save(totperim,file=here::here("data","NewlyCreatedData","TotalPerimetre.RData"))
```

# Evolution of size area of patches over time

```{r}

reg<-sort(unique(totsizearea$UMMPs))

#Filter landscape variable
totsizearea<-totsizearea %>% 
  dplyr::filter(Group!="MF" & Group !="JR" & Group != "PH2")

totshape<-totshape %>% 
  dplyr::filter(Group!="MF" & Group !="JR" & Group != "PH2")

totperim<-totperim %>% 
  dplyr::filter(Group!="MF" & Group !="JR" & Group != "PH2")

#get plots for each region
for (k in 1:length(reg)){

  #subset variable
xx<-totsizearea %>% 
  dplyr::filter(UMMPs==reg[k])
yy<-totshape %>% 
  dplyr::filter(UMMPs==reg[k])
zz<-totperim %>% 
  dplyr::filter(UMMPs==reg[k])

#plot variable by region
sizegg<-ggplot(xx,aes(x=Year,y=Size,group=Group)) +
  geom_line(aes(color=Group)) +
  facet_wrap(.~Group) +
  scale_colour_viridis_d(option="turbo")
  print(sizegg)
  
 perimgg<-ggplot(zz,aes(x=Year,y=Perimetre,group=Group)) +
  geom_line(aes(color=Group)) +
  facet_wrap(.~Group) +
  scale_colour_viridis_d(option="turbo")
  print(perimgg)
  
  shapegg<-ggplot(yy,aes(x=Year,y=Shape,group=Group)) +
  geom_line(aes(color=Group)) +
  facet_wrap(.~Group) +
  scale_colour_viridis_d(option="turbo")
  print(shapegg)
  
   setTxtProgressBar(progress_bar(length(reg)),k)  
   
  tiff(here::here("outputs","Size Area",paste0("Size_",reg[k],".tif")),
       res=300,
       width=2000,
       height=1500,
       units="px",
       compression="lzw")
  print(sizegg)
  dev.off()
  
    tiff(here::here("outputs","Shape",paste0("Shape_",reg[k],".tif")),
       res=300,
       width=2000,
       height=1500,
       units="px",
       compression="lzw")
  print(shapegg)
  dev.off()
  
    tiff(here::here("outputs","Perimeter",paste0("Perimeter_",reg[k],".tif")),
       res=300,
       width=2000,
       height=1500,
       units="px",
       compression="lzw")
  print(perimgg)
  dev.off()
  
  }
```
